Move well known lookup into a separate clas
